refactor: rename pipelines as workflows
