+++ backend/openedx_ai_extensions/api/v1/urls.py
@@ -4,10 +4,10 @@ Version 1 API URLs
 
 from django.urls import path
 
-from .pipelines.views import AIGenericWorkflowView
+from .workflows.views import AIGenericWorkflowView
 
 app_name = "v1"
 
 urlpatterns = [
-    path("workflows/", AIGenericWorkflowView.as_view(), name="ai_pipelines"),
+    path("workflows/", AIGenericWorkflowView.as_view(), name="ai_workflows"),
 ]
