+++ backend/openedx_ai_extensions/workflows/models.py
@@ -1,5 +1,5 @@
 """
-AI Workflow models for managing flexible AI pipeline execution
+AI Workflow models for managing flexible AI workflow execution
 """
 
 import logging
